Enhance SDK functionality and resolve naming conflicts
- Added new components for performance tracking, including `ABTestGenerationTracker` and `PerformanceGenerationTracker`, to improve A/B testing and generation performance monitoring.
- Introduced `BenchmarkMetricsCollector`, `MetricsAggregator`, and `ReportGenerator` for comprehensive benchmarking and performance reporting.
- Implemented `DownloadProgressAggregator` and `DownloadProgressTracker` to manage download progress effectively.
- Updated `ModelMetadata` with additional properties for enhanced model validation.
- Added `DownloadRetryStrategy` for improved error handling in download processes.
- Resolved naming conflicts by renaming files and classes for clarity and consistency.
- Removed obsolete `DownloadError` file to streamline error handling.
- Updated documentation to reflect new features and architectural changes.

+++ sdk/runanywhere-swift/Sources/RunAnywhere/Capabilities/Benchmarking/Services/ComparisonEngine.swift
@@ -74,9 +74,9 @@ public class ComparisonEngine {
         }
 
         // Tertiary metric: memory usage
-        if result1.memoryUsed < result2.memoryUsed * 0.8 {
+        if result1.memoryUsed < Int64(Double(result2.memoryUsed) * 0.8) {
             return 1
-        } else if result2.memoryUsed < result1.memoryUsed * 0.8 {
+        } else if result2.memoryUsed < Int64(Double(result1.memoryUsed) * 0.8) {
             return 2
         }
